Refactor: Manage Google Tag Manager & Google Analytics with next/third-parties
